refactor: Moved example messages to separate package
