+++ src/main/java/net/dancier/kikeriki/KikerikiApplicationConfiguration.java
@@ -1,5 +1,6 @@
 package net.dancier.kikeriki;
 
+import net.dancier.kikeriki.messages.Message;
 import org.apache.kafka.clients.consumer.Consumer;
 import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
