+++ src/main/java/net/dancier/kikeriki/MessageHandler.java
@@ -1,5 +1,9 @@
 package net.dancier.kikeriki;
 
+import net.dancier.kikeriki.messages.MessageBar;
+import net.dancier.kikeriki.messages.MessageFoo;
+
+
 public interface MessageHandler
 {
   void handleFoo(String key, MessageFoo foo);
